add backbutton to component, fixed styling
